requirements.txt
django>=5
djangorestframework
psycopg[binary]
Pillow

djangorestframework
djangorestframework-gis